create dynamic installation instructions
